function "executeOldDhcpAction" partially implemented.
When the MUD manager receives a request from an OLD device with an associated MUD URL, the MUD manager retrieves the MUD file and verifies if it is different from the version already available.
If so, the old firewall rules are deleted and the rules contained in the new MUD file are enforced.